+++ src/js/alerts-list.js
@@ -13,6 +13,3 @@ $( '.inline-edit-col-left, .inline-edit-col-right, #major-publishing-actions', '
 		$( this ).remove();
 	},
 );
-
-// This is done with JS instead of CSS to override the inline styles added by Select2's JS.
-$( '.select2-container', '.inline-edit-col' ).css( { width: '100%' } );