+++ ui/js/alerts-list.js
@@ -8,9 +8,6 @@
 				},
 			);
 
-			// This is done with JS instead of CSS to override the inline styles added by Select2's JS.
-			$( '.select2-container', '.inline-edit-col' ).css( { width: '100%' } );
-
 			// Re-enable the select all functionality
 			$( '.wp-list-table thead .check-column input[type="checkbox"]' ).on(
 				'click',